* Fixes from NicolaSix \ LambdaSix from the forum \ Github that addresses nulref spam in the console when accessing the part action window in KSP 1.12.1.
If you're still around Dmagic, these are edits to your .dll so i'm not happy about pushing it public, but it keeps the mod functional in the meantime :) Will test on my machine, Nicola is a avid contributer to BDB so I'm sure her work is fine, but will try bundle it with the stuff i never released for an official US2 release for 1.12.1. Give it a look over when you can and see what you think.
So yeah, decided to set up the project again after it got wiped with a OS issue. No lost data, just have to rebuild everything. Huge job but i think i'm up to it after all this time.

Hope you guys are well, if you still check this or get these updates.

Files taken from here,
https://github.com/LambdaSix/UniversalStorage2_1.12/releases/tag/1.8.1a